qnap_qsw: add entity classes
Reuses code where possible.
