+++ homeassistant/components/qnap_qsw/__init__.py
@@ -1,65 +1,17 @@
 """The QNAP QSW integration."""
 from __future__ import annotations
 
-from typing import Any
-
-from aioqsw.const import (
-    QSD_FIRMWARE,
-    QSD_FIRMWARE_INFO,
-    QSD_MAC,
-    QSD_PRODUCT,
-    QSD_SYSTEM_BOARD,
-)
 from aioqsw.localapi import ConnectionOptions, QnapQswApi
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_PASSWORD, CONF_URL, CONF_USERNAME, Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import aiohttp_client
-from homeassistant.helpers.device_registry import CONNECTION_NETWORK_MAC
-from homeassistant.helpers.entity import DeviceInfo
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .const import DOMAIN, MANUFACTURER
+from .const import DOMAIN
 from .coordinator import QswUpdateCoordinator
 
 PLATFORMS: list[Platform] = [Platform.BINARY_SENSOR, Platform.SENSOR]
-
-
-class QswEntity(CoordinatorEntity[QswUpdateCoordinator]):
-    """Define an QNAP QSW entity."""
-
-    def __init__(
-        self,
-        coordinator: QswUpdateCoordinator,
-        entry: ConfigEntry,
-    ) -> None:
-        """Initialize."""
-        super().__init__(coordinator)
-
-        self.product = self.get_device_value(QSD_SYSTEM_BOARD, QSD_PRODUCT)
-        self._attr_device_info = DeviceInfo(
-            configuration_url=entry.data[CONF_URL],
-            connections={
-                (
-                    CONNECTION_NETWORK_MAC,
-                    self.get_device_value(QSD_SYSTEM_BOARD, QSD_MAC),
-                )
-            },
-            manufacturer=MANUFACTURER,
-            model=self.product,
-            name=self.product,
-            sw_version=self.get_device_value(QSD_FIRMWARE_INFO, QSD_FIRMWARE),
-        )
-
-    def get_device_value(self, key: str, subkey: str) -> Any:
-        """Return device value by key."""
-        value = None
-        if key in self.coordinator.data:
-            data = self.coordinator.data[key]
-            if subkey in data:
-                value = data[subkey]
-        return value
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
