+++ homeassistant/components/qnap_qsw/binary_sensor.py
@@ -16,17 +16,18 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity import EntityCategory
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import QswEntity
 from .const import ATTR_MESSAGE, DOMAIN
 from .coordinator import QswUpdateCoordinator
+from .entity import QswEntityDescription, QswSensorEntity
 
 
 @dataclass
-class QswBinarySensorEntityDescription(BinarySensorEntityDescription):
+class QswBinarySensorEntityDescription(
+    BinarySensorEntityDescription, QswEntityDescription
+):
     """A class that describes QNAP QSW binary sensor entities."""
 
     attributes: dict[str, list[str]] | None = None
-    subkey: str = ""
 
 
 BINARY_SENSOR_TYPES: Final[tuple[QswBinarySensorEntityDescription, ...]] = (
@@ -58,7 +59,7 @@ async def async_setup_entry(
     )
 
 
-class QswBinarySensor(QswEntity, BinarySensorEntity):
+class QswBinarySensor(QswSensorEntity, BinarySensorEntity):
     """Define a QNAP QSW binary sensor."""
 
     entity_description: QswBinarySensorEntityDescription
@@ -79,20 +80,9 @@ class QswBinarySensor(QswEntity, BinarySensorEntity):
         self._async_update_attrs()
 
     @callback
-    def _handle_coordinator_update(self) -> None:
-        """Update attributes when the coordinator updates."""
-        self._async_update_attrs()
-        super()._handle_coordinator_update()
-
-    @callback
     def _async_update_attrs(self) -> None:
         """Update binary sensor attributes."""
         self._attr_is_on = self.get_device_value(
             self.entity_description.key, self.entity_description.subkey
         )
-
-        if self.entity_description.attributes:
-            self._attr_extra_state_attributes = {
-                key: self.get_device_value(val[0], val[1])
-                for key, val in self.entity_description.attributes.items()
-            }
+        super()._async_update_attrs()
